* Use 'long long' to print sector sizes, to avoid overflowing 'long'.
git-svn-id: svn://svn.debian.org/svn/d-i/trunk@36044 48c42b26-1dd6-0310-b98f-a58d8bce7237

+++ packages/autopartkit/distribute.c
@@ -154,8 +154,8 @@ distribute_partitions(struct disk_info_t diskinfo[],
 		    reqs[j].max_blk = maxmax_blk;
 		span = reqs[j].max_blk - reqs[j].min_blk;
 		total_wanted += span;
-		autopartkit_log(0, "Adding %ld to total_wanted, now %ld\n",
-				(long)span, (long)total_wanted);
+		autopartkit_log(0, "Adding %lld to total_wanted, now %lld\n",
+				(long long)span, (long long)total_wanted);
 	    }
 	}
         if (total_wanted == 0)
